Refactor pages
Move to separate file and inline the `Config` struct.

Also add `prettier` processing to Makefile.
