+++ docs/index.html
@@ -1,20 +1,23 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <title>SSSSG! 😎</title>
-        <meta name="description" content="Super Simple Static Site Generator is great for making static pages." />
-    </head>
-    <body>
-      <h1>SSSSG! 😎</h1>
-<p>The Super Simple Static Site Generator makes it incredibly easy to build static sites.</p>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>SSSSG! 😎</title>
+    <meta
+      name="description"
+      content="Super Simple Static Site Generator is great for making static pages."
+    />
+  </head>
+  <body>
+    <h1>SSSSG! 😎</h1>
+    <p>
+      The Super Simple Static Site Generator makes it incredibly easy to build
+      static sites.
+    </p>
 
-<p>It has composite components.</p>
+    <p>It has composite components.</p>
 
-<p>This is the nested component.</p>
-
-
-
-    </body>
+    <p>This is the nested component.</p>
+  </body>
 </html>
