+++ layouts/default.html
@@ -3,8 +3,8 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <title>{{.Config.Meta.Title}}</title>
-        <meta name="description" content="{{.Config.Meta.Description}}" />
+        <title>{{.Meta.Title}}</title>
+        <meta name="description" content="{{.Meta.Description}}" />
     </head>
     <body>
       {{.Body}}
