+++ ssssg.go
@@ -2,30 +2,13 @@ package ssssg
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"os/exec"
 	"path"
 	"strings"
 	"text/template"
-
-	yaml "gopkg.in/yaml.v2"
 )
-
-type Page struct {
-	Path   string
-	Config Config
-	Body   string
-	Layout string
-}
-
-type Config struct {
-	Meta struct {
-		Title       string
-		Description string
-	}
-}
 
 type BuildOptions struct {
 	BuildDir      string
@@ -43,7 +26,7 @@ func Build(options BuildOptions) error {
 		return fmt.Errorf("could not copy static files")
 	}
 
-	pages, err := exploreDir(options.PagesDir)
+	pages, err := readPages(options.PagesDir)
 	if err != nil {
 		return err
 	}
@@ -72,10 +55,6 @@ func Build(options BuildOptions) error {
 		}
 		page.Body = content.String()
 
-		if page.Layout == "" {
-			page.Layout = "default.html"
-		}
-
 		outputPath := path.Join(options.BuildDir, strings.TrimSuffix(strings.TrimPrefix(page.Path, options.PagesDir+"/"), ".yaml")) + ".html"
 
 		if err := os.MkdirAll(path.Dir(outputPath), 0766); err != nil {
@@ -92,49 +71,10 @@ func Build(options BuildOptions) error {
 			return err
 		}
 
-		if err := t.ExecuteTemplate(output, page.Layout, page); err != nil {
+		if err := t.ExecuteTemplate(output, "default.html", page); err != nil {
 			return err
 		}
 	}
 
 	return nil
-}
-
-func exploreDir(dir string) ([]Page, error) {
-	var pages []Page
-	entries, err := ioutil.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-	for _, entry := range entries {
-		if entry.IsDir() {
-			subPages, err := exploreDir(path.Join(dir, entry.Name()))
-			if err != nil {
-				return nil, err
-			}
-			pages = append(pages, subPages...)
-			continue
-		}
-		if !strings.HasSuffix(entry.Name(), ".yaml") {
-			continue
-		}
-
-		configContent, err := ioutil.ReadFile(path.Join(dir, entry.Name()))
-		if err != nil {
-			return nil, err
-		}
-		var config Config
-		if err := yaml.Unmarshal(configContent, &config); err != nil {
-			return nil, err
-		}
-
-		body, err := ioutil.ReadFile(path.Join(dir, strings.TrimSuffix(entry.Name(), ".yaml")+".html"))
-		page := Page{
-			Path:   path.Join(dir, entry.Name()),
-			Config: config,
-			Body:   string(body),
-		}
-		pages = append(pages, page)
-	}
-	return pages, nil
 }
